@andy-js 2017-10-01
You forgot to add a mapping for taskqid_t to sys/zfs_context.h.