@szaydel 2017-10-03
Perhaps a small knit, maybe not even a knit. I noticed that `int ret = ENOENT;` was added in `usr/src/lib/libzfs/common/libzfs_dataset.c` at line 844, but it does not appear as though this `ret` variable is used consistently. Maybe it should be initialized to `0` (it is reassigned later to `0`, seemingly) and then set accordingly from return of call, like one to `getmntany` at line 854. 
Would be good I suppose if it was returned consistently, as opposed to being used in some cases.