@andy-js 2017-10-24
@prakashsurya Do you think it would make sense to split the changes to the VFS code out into a separate issue?