@andy-js 2017-10-25
I took at stab at updating the changeset to use libfakekernel instead:
http://cr.illumos.org/~webrev/andy_js/8115/

Apart from some weirdness with sys/cmn_err.h conflicting with stdio.h it was straightforward.