@andy-js 2017-10-25
Well that depends on whether or not they're pulling in both libzfs and libzpool.  From what I can see most things (like the zfs and zpool commands) only pull in libzfs, so they should be okay.

I have no problem with changing libzpool to use libfakekernel.  I chose not to go down that route simply because I wanted to keep the diff small, but I think it's probably the right thing to do.