@prakashsurya 2017-10-26
I like where this is going.

IMO, I think we should split the taskq changes out from this change (as suggested), do what's needed to get libzfs (and maybe libzpool also) using libfakekernel, and then apply what's left of this change on top of the taskq changes.

This way, there's a clear separation between the taskq changes that shouldn't have any "external" impact to the CLI tools and/or library consumers (right?), and a separate patch to implement the actual "feature" of this change using the libfakekernel taskq implementation.

@andy-js, you've pretty much done this already, so I presume you're on board with this; @gwr does this sound good to you too?