@andy-js 2017-11-01
Sounds good to me.  I'll look at updating libzpool to use libfakekernel.