@andy-js 2017-11-06
I introduced _TASKQUSER so that we don't need to build libzfs in fakekernel context, which reduces the size of the diff a bit.