@ikozhukhov 2017-11-27
what is it status of this PR?
maybe we can update it later/by next changes for taskq?
i'd like to see it integrated